+++ packages/addons/drizzle/index.ts
@@ -74,7 +74,7 @@ export default defineAddon({
 	run: ({ sv, typescript, options, kit }) => {
 		const ext = typescript ? 'ts' : 'js';
 
-		sv.dependency('drizzle-orm', '^0.38.4');
+		sv.dependency('drizzle-orm', '^0.40.0');
 		sv.devDependency('drizzle-kit', '^0.30.2');
 
 		// MySQL
@@ -116,6 +116,10 @@ export default defineAddon({
 					dbSpecificContent = `
                       MYSQL_ROOT_PASSWORD: ${PASSWORD}
                       MYSQL_DATABASE: ${DB_NAME}
+                    volumes:
+                      - mysqldata:/var/lib/mysql
+                volumes:
+                  mysqldata:
                 `;
 				}
 				if (options.postgresql === 'postgres.js') {
@@ -123,6 +127,10 @@ export default defineAddon({
                       POSTGRES_USER: ${USER}
                       POSTGRES_PASSWORD: ${PASSWORD}
                       POSTGRES_DB: ${DB_NAME}
+                    volumes:
+                      - pgdata:/var/lib/postgresql/data
+                volumes:
+                  pgdata:
                 `;
 				}
 
@@ -191,8 +199,8 @@ export default defineAddon({
 					url: common.expressionFromString('process.env.DATABASE_URL'),
 					authToken
 				}),
-				verbose: { type: 'BooleanLiteral', value: true },
-				strict: { type: 'BooleanLiteral', value: true }
+				verbose: { type: 'Literal', value: true },
+				strict: { type: 'Literal', value: true }
 			});
 
 			const dialect = options.sqlite === 'turso' ? 'turso' : options.database;
@@ -261,6 +269,7 @@ export default defineAddon({
 			const { ast, generateCode } = parseScript(content);
 
 			imports.addNamed(ast, '$env/dynamic/private', { env: 'env' });
+			imports.addNamespace(ast, './schema', 'schema');
 
 			// env var checks
 			const dbURLCheck = common.statementFromString(
@@ -296,19 +305,13 @@ export default defineAddon({
 				}
 			}
 			// MySQL
-			if (options.mysql === 'mysql2') {
+			if (options.mysql === 'mysql2' || options.mysql === 'planetscale') {
 				imports.addDefault(ast, 'mysql2/promise', 'mysql');
 				imports.addNamed(ast, 'drizzle-orm/mysql2', { drizzle: 'drizzle' });
 
 				clientExpression = common.expressionFromString(
 					'await mysql.createConnection(env.DATABASE_URL)'
 				);
-			}
-			if (options.mysql === 'planetscale') {
-				imports.addNamed(ast, '@planetscale/database', { Client: 'Client' });
-				imports.addNamed(ast, 'drizzle-orm/planetscale-serverless', { drizzle: 'drizzle' });
-
-				clientExpression = common.expressionFromString('new Client({ url: env.DATABASE_URL })');
 			}
 			// PostgreSQL
 			if (options.postgresql === 'neon') {
@@ -328,7 +331,20 @@ export default defineAddon({
 			const clientIdentifier = variables.declaration(ast, 'const', 'client', clientExpression);
 			common.addStatement(ast, clientIdentifier);
 
+			// create drizzle function call
 			const drizzleCall = functions.callByIdentifier('drizzle', ['client']);
+
+			// add schema to support `db.query`
+			const paramObject = object.create({
+				schema: variables.identifier('schema')
+			});
+			if (options.database === 'mysql') {
+				const mode = options.mysql === 'planetscale' ? 'planetscale' : 'default';
+				object.property(paramObject, 'mode', common.createLiteral(mode));
+			}
+			drizzleCall.arguments.push(paramObject);
+
+			// create `db` export
 			const db = variables.declaration(ast, 'const', 'db', drizzleCall);
 			exports.namedExport(ast, 'db', db);
 
@@ -342,6 +358,10 @@ export default defineAddon({
 		if (options.docker) {
 			steps.push(
 				`Run ${highlighter.command(`${packageManager} run db:start`)} to start the docker container`
+			);
+		} else {
+			steps.push(
+				`Check ${highlighter.env('DATABASE_URL')} in ${highlighter.path('.env')} and adjust it to your needs`
 			);
 		}
 		steps.push(