+++ packages/cli/utils/package-manager.ts
@@ -8,7 +8,7 @@ import {
 	AGENTS,
 	COMMANDS,
 	constructCommand,
-	detectSync,
+	detect,
 	type AgentName
 } from 'package-manager-detector';
 import { parseJson } from '@sveltejs/cli-core/parsers';
@@ -19,8 +19,12 @@ agentOptions.unshift({ label: 'None', value: undefined });
 
 type PackageManagerOptions = Array<{ value: AgentName | undefined; label: AgentName | 'None' }>;
 export async function packageManagerPrompt(cwd: string): Promise<AgentName | undefined> {
-	const detected = detectSync({ cwd });
+	const detected = await detect({ cwd });
 	const agent = detected?.name ?? getUserAgent();
+
+	// If we are in a non interactive environment just go with the detected package manager.
+	// There is no need to prompt in that case.
+	if (!process.stdout.isTTY) return agent;
 
 	const pm = await p.select({
 		message: 'Which package manager do you want to install dependencies with?',