+++ packages/core/tests/js/arrays/object-array/output.ts
@@ -1,8 +1,0 @@
-const array = [
-	{
-		test: true
-	},
-	{
-		test2: 'string'
-	}
-];