+++ packages/core/tests/js/arrays/push-object-array/output.ts
@@ -1,0 +1,1 @@
+const array = [{ test: true }, { test2: 'string' }];