+++ packages/core/tests/js/functions/function-call-by-identifier/output.ts
@@ -3,4 +3,5 @@ function foo(bar: string) {
 }
 
 const a = 'bar';
+
 foo(a);