+++ packages/core/tests/js/imports/empty-import/run.ts
@@ -1,7 +1,6 @@
-import { imports } from '@sveltejs/cli-core/js';
-import type { ScriptFileEditor } from '@sveltejs/cli-core';
+import { imports, type AstTypes } from '@sveltejs/cli-core/js';
 
-export function run({ ast }: ScriptFileEditor<any>): void {
+export function run(ast: AstTypes.Program): void {
 	imports.addEmpty(ast, './relativ/file.css');
 
 	// allow importing from npm packages