+++ packages/core/tests/js/variables/identifier/run.ts
@@ -1,7 +1,6 @@
-import { variables } from '@sveltejs/cli-core/js';
-import type { ScriptFileEditor } from '@sveltejs/cli-core';
+import { variables, type AstTypes } from '@sveltejs/cli-core/js';
 
-export function run({ ast }: ScriptFileEditor<any>): void {
+export function run(ast: AstTypes.Program): void {
 	const barVariable = variables.declaration(ast, 'const', 'bar', variables.identifier('foo'));
 	ast.body.push(barVariable);
 }